requirements.txt
altgraph==0.17.3
greenlet==2.0.2
install==1.3.5
macholib==1.16.2
numpy==1.25.0
pandas==2.0.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.17
typing_extensions==4.6.3
tzdata==2023.3
